--- jsconfig.json
@@ -1,21 +1,0 @@
-{
-    "compilerOptions": {
-        "target": "ES6",
-        "module": "commonjs",
-        "allowSyntheticDefaultImports": true,
-        "baseUrl": "./",
-        "paths": {
-          "@components/*": ["src/components/*"],
-          "@utils/*": ["src/api/utils/*"],
-          "@const/*": ["src/api/const/*"],
-          "@api/*": ["src/api/*"],
-          "@layouts/*": ["src/layouts/*"],
-          "@pages/*": ["src/pages/*"],
-          "@properties/*": ["src/properties/*"],
-          "@/*": ["src/*"],
-        }
-    },
-    "exclude": [
-        "node_modules"
-    ]
-}
